requirements.txt
absl-py==2.4.0
anyio==4.12.1
asgiref==3.9.1
astunparse==1.6.3
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.4
django-cors-headers==4.9.0
django-firebase-auth==1.2.4
djangorestframework==3.16.0
firebase_admin==7.2.0
flatbuffers==25.12.19
gast==0.7.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.11.0
requests==2.32.5
rich==15.0.0
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.3
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
wheel==0.46.3
wrapt==2.1.2
